Move tests for `rangeTo` into separate file
